refactor: refactored background_tasks & indexing_tasks
